Begin some minor fixes
Saving program space on ESP32 by condensing JS functions.
Found a bug while saving smoke settings from the Attenuator--the values were changed to uint32_t though the max stored is 60000 which can fit inside of a uint16_t. Somehow the values sent back from the ESP32 were only uint8_t but after multiplying out to convert seconds to milliseconds the values were getting "corrupted".

+++ source/NeutronaWand/Configuration.h
@@ -198,12 +198,13 @@ bool b_overheat_level_5 = true;
  * Time in milliseconds for when overheating will initiate if enabled for that power level.
  * Overheat only happens if enabled for that power level (see above).
  * Example: 60000 = (60 seconds)
+ * Max is 60000 to keep within the uint16 integer type.
  */
-uint32_t i_ms_overheat_initiate_level_1 = 60000;
-uint32_t i_ms_overheat_initiate_level_2 = 50000;
-uint32_t i_ms_overheat_initiate_level_3 = 40000;
-uint32_t i_ms_overheat_initiate_level_4 = 35000;
-uint32_t i_ms_overheat_initiate_level_5 = 30000;
+uint16_t i_ms_overheat_initiate_level_1 = 60000;
+uint16_t i_ms_overheat_initiate_level_2 = 50000;
+uint16_t i_ms_overheat_initiate_level_3 = 40000;
+uint16_t i_ms_overheat_initiate_level_4 = 35000;
+uint16_t i_ms_overheat_initiate_level_5 = 30000;
 
 /*
  * Set to false to disable the onboard amplifier on the WAV Trigger.
